fix(major): address 44 Phase 2 major correctness issues (M1-M44)
Key fixes across all modules:

ANALYSIS & CHANGE DETECTION:
- M2: bytecodeChangedMethodKeys populated on first run when prev store is empty
- M3: BytecodeHashStore visitLookupSwitchInsn hashes label count for consistency
- M4: StaticCallGraphAnalyzer addCallEdge BFS no longer stops at first ancestor
- M6/M7: SourceFileModel text-block line tracking and triple-quote emission fixed
- M8: StructuralDiff.readFilesFromGit uses OID-based matching instead of positional
- M36: getChangedMethodKeys returns empty when prev has classes but no methods
- M37: MethodHashStore registers KotlinFooKt class name variant for .kt files

SCORING:
- M9: Backward-compat TestScorer constructors pass normalized weightedDepOverlap
- M10: killMultiplier() extracted to shared method used by score() and explain()

JUNIT EXTENSION:
- M11: finishedNormally=true set after all IO and map clears complete
- M12: buildRunRecord calls synchronized on executionOrder
- M13: PriorityClassOrderer captures originalCopy before clear() for safe recovery
- M14: PriorityMethodOrderer replaces 6 static volatiles with AtomicReference
- M15: scoreIndexMap keyed by name(paramTypes) to distinguish overloads

TEST DETECTION / WORKFLOW:
- M16: TddEnforcementCore skips rename detection for nested classes
- M17: Class.forName uses thread context classloader
- M19: TestSelector.selectTopN only increments counted on successful Set.add()
- M20: TieredTestSelector budget uses double arithmetic to prevent Long overflow
- M21: LearnWorkflow logs IOException before falling back
- M22: AutoWorkflow alwaysLearn pre-aggregation runs before mode resolution
- M23: DependencyMap.filterForModule applies filter to methodDependencies
- M24: DependencyMap.mergeWith calls invalidateCaches() instead of direct nulls
- M43: ChangeAnalysis module filter no longer discards valid filtered result
- M44: ChangeAnalysis.joinOrRethrow re-throws Error before IOException wrapping

PERSISTENCE / SERIALIZATION:
- M25: PersistenceSupport.pruneLocks uses removeIf() for atomic per-entry removal
- M26: StateSerializer detects LZ4 by correct magic bytes (4C5A3442/04224D18)
- M27: StateMigrations.migrate uses copy-on-write for immutable map safety

AGENT:
- M28: AsmClassTransformer re-throws Error before catching Exception
- M29: IntelligentClassFilter only decrements cache size on genuine eviction
- M30: IndexCollectorClient/Server use 4-byte int length prefix (was 2-byte short)

OPERATIONS:
- M31: DetectDependenciesOperation uses JSONParser instead of string-scanning
- M33: DeltaDebugging logs warning when budget exhausted
- M35: FailureHistoryTracker comment corrects misleading "atomic snapshot" claim
- M38: CombinedAdaptiveAlgorithm includes ISOLATE_PAIR in multi-polluter bypass
- M39: ReactorReorderer.buildPredecessors uses getOrDefault to avoid NPE
- M40: CollectorLifecycleParticipant removes session key after drain loop
- M41: TestOrderPlugin uses jvmArgs() instead of System.setProperty()
- M42: GradleTestRunner.writeInitScript synchronized

Also A5: DependencyMap.dependencies uses TreeMap for deterministic serialization

+++ test-order-agent/src/main/java/me/bechberger/testorder/agent/runtime/IndexCollectorClient.java
@@ -31,7 +31,7 @@ import java.util.Set;
  * [1 byte]  version: 2
  * [4 bytes] test-class tracker count
  * per tracker:
- *   [2 bytes] key length (UTF-8)
+ *   [4 bytes] key length (UTF-8)
  *   [N bytes] key (test class name)
  *   [4 bytes] classWords length (number of longs)
  *   [M×8 bytes] classWords data
@@ -39,7 +39,7 @@ import java.util.Set;
  *   [K×8 bytes] memberWords data
  * [4 bytes] method tracker count
  * per tracker:
- *   [2 bytes] key length (UTF-8)
+ *   [4 bytes] key length (UTF-8)
  *   [N bytes] key (className#methodName)
  *   [4 bytes] classWords length
  *   [M×8 bytes] classWords data
@@ -233,7 +233,7 @@ public final class IndexCollectorClient {
 
 	private static void writeString(DataOutputStream out, String s) throws IOException {
 		byte[] bytes = s.getBytes(StandardCharsets.UTF_8);
-		out.writeShort(bytes.length);
+		out.writeInt(bytes.length);
 		out.write(bytes);
 	}
 }
